more localization and validation

+++ src/main/java/by/guzov/finaltask/command/CommandDeleteRequest.java
@@ -3,10 +3,12 @@ package by.guzov.finaltask.command;
 import by.guzov.finaltask.domain.Request;
 import by.guzov.finaltask.domain.User;
 import by.guzov.finaltask.dto.ResponseContent;
+import by.guzov.finaltask.i18n.MessageLocalizer;
 import by.guzov.finaltask.service.RequestService;
 import by.guzov.finaltask.service.ServiceException;
 import by.guzov.finaltask.service.ServiceFactory;
 import by.guzov.finaltask.util.AppConstants;
+import by.guzov.finaltask.validation.StringValidator;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -15,17 +17,22 @@ public class CommandDeleteRequest implements Command {
     public ResponseContent execute(HttpServletRequest request) {
         try {
             User sessionUser = ((User) request.getSession().getAttribute(AppConstants.SESSION_USER));
-            int requestId = Integer.parseInt(request.getParameter(AppConstants.ID));
+            String id = request.getParameter(AppConstants.ID);
+            if (!StringValidator.isValid(id, 1, 9, StringValidator.NUMBER_PATTERN)) {
+                return ResponseUtil.toCommandWithError(request,
+                        CommandType.SHOW_EMPTY_PAGE, "field.id" + MessageLocalizer.DELIMITER + "error.invalid_base");
+            }
+            int requestId = Integer.parseInt(id);
             RequestService requestService = ServiceFactory.getInstance().getRequestService();
             Request current = requestService.getById(requestId);
             if (sessionUser.getRole().equals(AppConstants.ADMIN) || sessionUser.getId() == current.getUserId()) {
                 requestService.delete(current);
                 return ResponseUtil.redirectWIthSuccess(request, CommandType.SHOW_REQUEST_LIST.name());
             } else {
-                return ResponseUtil.toCommandWithError(request, CommandType.SHOW_EMPTY_PAGE, "you are forbidden to do this");
+                return ResponseUtil.toCommandWithError(request, CommandType.SHOW_EMPTY_PAGE, "error.forbidden");
             }
         } catch (ServiceException e) {
-            return ResponseUtil.redirectTo(request, CommandType.SHOW_EMPTY_PAGE + "error_message=server_error");
+            return ResponseUtil.redirectTo(request, CommandType.SHOW_EMPTY_PAGE + "error_message=server%error");
         }
     }
 }
